+++ crates/hyperqueue/src/server/client/submit.rs
@@ -470,6 +470,7 @@ mod tests {
                 env: Default::default(),
                 stdout: Default::default(),
                 stderr: Default::default(),
+                stdin: vec![],
                 cwd: Default::default(),
             },
             resources: ResourceRequest {